Move observable utilities into `utilities/observables`
