+++ src/__tests__/local-state/export.ts
@@ -1,7 +1,7 @@
 import gql from 'graphql-tag';
 import { print } from 'graphql/language/printer';
 
-import { Observable } from '../../util/Observable';
+import { Observable } from '../../utilities/observables/Observable';
 import { ApolloLink } from '../../link/core/ApolloLink';
 import { ApolloClient } from '../..';
 import { InMemoryCache } from '../../cache/inmemory/inMemoryCache';
