+++ src/__tests__/local-state/subscriptions.ts
@@ -1,6 +1,6 @@
 import gql from 'graphql-tag';
 
-import { Observable } from '../../util/Observable';
+import { Observable } from '../../utilities/observables/Observable';
 import { ApolloLink } from '../../link/core/ApolloLink';
 import { ApolloClient } from '../..';
 import { InMemoryCache } from '../../cache/inmemory/inMemoryCache';
