+++ src/__tests__/optimistic.ts
@@ -5,7 +5,7 @@ import gql from 'graphql-tag';
 
 import { mockSingleLink } from '../__mocks__/mockLinks';
 import { MutationQueryReducersMap } from '../core/types';
-import { Subscription } from '../util/Observable';
+import { Subscription } from '../utilities/observables/Observable';
 import { ApolloClient } from '../';
 import { addTypenameToDocument } from '../utilities/graphql/transform';
 import { stripSymbols } from './utils/stripSymbols';
