+++ src/core/__tests__/ObservableQuery.ts
@@ -1,7 +1,7 @@
 import gql from 'graphql-tag';
 import { GraphQLError } from 'graphql';
 
-import { Observable } from '../../util/Observable';
+import { Observable } from '../../utilities/observables/Observable';
 import { ApolloLink } from '../../link/core/ApolloLink';
 import { InMemoryCache } from '../../cache/inmemory/inMemoryCache';
 import mockQueryManager from '../../__mocks__/mockQueryManager';
