+++ src/link/core/__tests__/ApolloLink.ts
@@ -1,7 +1,7 @@
 import gql from 'graphql-tag';
 import { print } from 'graphql/language/printer';
 
-import { Observable } from '../../../util/Observable';
+import { Observable } from '../../../utilities/observables/Observable';
 import { FetchResult, Operation, NextLink, GraphQLRequest } from '../types';
 import { ApolloLink } from '../ApolloLink';
