+++ src/link/core/types.ts
@@ -2,7 +2,7 @@ import { DocumentNode } from 'graphql/language/ast';
 import { ExecutionResult } from 'graphql/execution/execute';
 export { ExecutionResult, DocumentNode };
 
-import { Observable } from '../../util/Observable';
+import { Observable } from '../../utilities/observables/Observable';
 
 export interface GraphQLRequest {
   query: DocumentNode;
