+++ src/link/utils/fromError.ts
@@ -1,4 +1,4 @@
-import { Observable } from '../../util/Observable';
+import { Observable } from '../../utilities/observables/Observable';
 
 export function fromError<T>(errorValue: any): Observable<T> {
   return new Observable<T>(observer => {
