+++ src/link/utils/fromPromise.ts
@@ -1,4 +1,4 @@
-import { Observable } from '../../util/Observable';
+import { Observable } from '../../utilities/observables/Observable';
 
 export function fromPromise<T>(promise: Promise<T>): Observable<T> {
   return new Observable<T>(observer => {
