+++ src/react/hooks/__tests__/useQuery.test.tsx
@@ -3,7 +3,7 @@ import { DocumentNode, GraphQLError } from 'graphql';
 import gql from 'graphql-tag';
 import { render, cleanup, wait } from '@testing-library/react';
 
-import { Observable } from '../../../util/Observable';
+import { Observable } from '../../../utilities/observables/Observable';
 import { ApolloLink } from '../../../link/core/ApolloLink';
 import { MockedProvider, MockLink } from '../../testing';
 import { ApolloClient } from '../../../ApolloClient';
