+++ src/react/types/types.ts
@@ -1,7 +1,7 @@
 import { ReactNode } from 'react';
 import { DocumentNode } from 'graphql';
 
-import { Observable } from '../../util/Observable';
+import { Observable } from '../../utilities/observables/Observable';
 import { FetchResult } from '../../link/core/types';
 import { ApolloClient } from '../../ApolloClient';
 import {
